requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==3.6.4.0
celery==5.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Click==7.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-appconf==1.0.5
django-compressor==4.4
django-sass-processor==1.3
djangorestframework==3.14.0
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
kombu==5.3.3
libsass==0.22.0
mccabe==0.7.0
packaging==23.2
prompt-toolkit==3.0.41
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.10